+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [FEATURE] Add watchOS and visionOS support. See [#2817][]
 - [FEATURE] Add support for the FedRAMP-compatible `fed2.ddog-gov.com` site. See [#2827][]
 - [IMPROVEMENT] Rename RUM Operations APIs. See [#2802][]
+- [FIX] Propagate native `anonymous_id` to WebView RUM and Log events. See [#2847][]
 
 # 3.9.1 / 15-04-2026
 
@@ -1120,6 +1121,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#2840]: https://github.com/DataDog/dd-sdk-ios/pull/2840
 [#2817]: https://github.com/DataDog/dd-sdk-ios/pull/2817
 [#2827]: https://github.com/DataDog/dd-sdk-ios/pull/2827
+[#2847]: https://github.com/DataDog/dd-sdk-ios/pull/2847
 
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin